==> perception/1_ready_for_R.ipynb <==
# Prep data file for analysis in R

import pandas as pd
import numpy as np

First get the list of partIDs and their language profiles.

parts = pd.read_csv("../blp/data/parts.csv")
parts['partID'] = parts['partID'].map(str.lower)
trilingual_list = parts[parts['lang_profile']!="mono"].copy().partID.values.tolist()
part_list = parts['partID'].values.tolist()

Now read in the data from each language's perception task. Drop the first 9828 rows from the Spanish task which include pre-experiment testing. Also drop any rows with a partID not in the df `parts`.

spa = pd.read_csv("data/spanish.csv")
spa = spa.iloc[9828:]
spa['partID'] = spa['partID'].map(str.lower)
# should be 50
spa.partID.nunique()

spa_parts = spa.partID.values.tolist()
print("Trilingual participants that don't show up in Spanish perception data:")
print(list(set(trilingual_list) - set(spa_parts)))
print("Spanish perception data participants that don't show up in trilingual part list:")
print(list(set(spa_parts) - set(trilingual_list)))

Fix.

# aei079 was given ID anm099 by accident
spa.loc[spa['partID']=="anm099",'partID'] = 'aei079'
# vrc053 was given ID oec103 by accident
spa.loc[spa['partID']=="oec103",'partID'] = 'vrc053'
# usm074 had incorrect url
spa.loc[((spa["partID"] == "bad_value") & (spa["subjID"] == "2642c56c09d355")),
        "partID"] = "usm074"

spa = spa[spa['partID'].isin(part_list)].copy()
spa.reset_index(inplace = True, drop = True)
# should be 50
spa['partID'].nunique()

Some participants could not complete the third trial and had to restart the experiment, so they show as having two rounds of trial one.

spa['block'] = spa.groupby(['partID', 'trial'], sort = False).cumcount()
spa.head()

Similarly, for Catalan, remove the first 2826 rows.

cat = pd.read_csv("data/catalan.csv")
cat = cat.iloc[2825:]
cat = cat.rename(columns={'SubjID': 'subjID'})
cat['partID'] = cat['partID'].map(str.lower)
cat.partID.nunique()

cat_parts = cat.partID.values.tolist()
print("Trilingual participants that don't show up in Catalan perception data:")
print(list(set(trilingual_list) - set(cat_parts)))
print("Catalan perception data participants that don't show up in trilingual part list:")
print(list(set(cat_parts) - set(trilingual_list)))

Fix.

# fixes ama302
cat.partID = cat.partID.str.replace('aoa302','ama302')
# fixes rre101
cat.loc[((cat["partID"] == "bad_value") & (cat["subjID"] == "263ff9a086a788")),
        "partID"] = "rre101"
# fixes ark038 -- next below lic063 in spreadsheet, wrong link was copied and used
cat.loc[cat['partID']=="ark038",'partID'] = 'lic063'

cat = cat[cat['partID'].isin(part_list)].copy()
cat.reset_index(inplace = True, drop = True)
# should be 50
cat['partID'].nunique()

Some participants could not complete the third trial and had to restart the experiment, so they show as having two rounds of trial one.

cat['block'] = cat.groupby(['partID', 'trial'], sort = False).cumcount()
cat.head()

The number of columns was changed for the English task partway through data collection. Also, drop the first 1100 rows which were testing before the experiment began.

eng = pd.read_csv('data/english.csv', names = ['subjID', 'partID', 'list', 'block', 'trial',
                                               'file1', 'filedur', 'loadtime', 'status',
                                               'response', 'rt'])
eng = eng.iloc[1101:]
eng['partID'] = eng['partID'].map(str.lower)
eng.partID.nunique()

eng_parts = eng.partID.values.tolist()
print("Participants that don't show up in English perception data:")
print(list(set(part_list) - set(eng_parts)))
print("English perception data participants that don't show up in part list:")
print(list(set(eng_parts) - set(part_list)))

eng = eng[eng['partID'].isin(part_list)].copy()
eng.reset_index(inplace = True, drop = True)
# should be 70
eng['partID'].nunique()

The English experiment also had an issue with the `block` column. Create a repeating counter only for the main trials (i.e., not the practice trials).

eng = eng[~eng['file1'].str.contains("practice")].copy()
eng['block'] = eng.groupby(['partID', 'trial'], sort = False).cumcount()
eng.reset_index(inplace = True, drop = True)
eng.head()

# eng_practice = eng[eng['file1'].str.contains("practice")].copy()
# eng_practice['block'] = 'practice'
# eng = pd.concat([eng_main, eng_practice], axis = 0)
# eng.sort_index(inplace = True)
# eng.head()

We want to remove responses that are not 'OK', code responses as 'pen' for penultimate stress and 'ult' for ultimate stress, create a new column with just the filename, then new columns listing the step of each measure. Then, each of these columns needs to be centered, where 1 is converted to -2, 2 is converted to -1, 3 is converted to 0, 4 is converted to 1, and 5 is converted to 2. Additionally, when values of 0 should be converted to 3 in the raw measures (since this is the neutral midpoint) and kept as 0 in the centered column.

spa["response_binary"] = spa["response"].map({'pondis': "pen", 'pondís': "ult"})
cat["response_binary"] = cat["response"].map({'pondis': "pen", 'pondís': "ult"})
eng["response_binary"] = eng["response"].map({'PONdiss': 'pen', 'ponDISS': "ult"})

print("before spa: ",len(spa),"\nbefore cat: ", len(cat), "\nbefore eng: ", len(eng))
spa = spa[spa["status"]=="OK"].copy()
cat = cat[cat["status"]=="OK"].copy()
eng = eng[eng["status"]=="OK"].copy()
print("after spa: ",len(spa),"\nafter cat: ", len(cat), "\nafter eng: ", len(eng))

# account for practice trials in English task
spa['filename'] = spa['file1'].str.split('spanish/',expand=True)[1]
cat['filename'] = cat['file1'].str.split('catalan/',expand=True)[1]
eng['filename'] = eng['file1'].str.split('english/',expand=True)[1]#apply(lambda x: x.split("english/")[1] if "english" in x else "0")

spa['formant'] = spa['filename'].apply(lambda x: float(x.split("F")[1][0]) if "F" in x else 3)
spa['spectilt'] = spa['filename'].apply(lambda x: float(x.split("st")[1][0]) if "st" in x else 3)
spa['duration'] = spa['filename'].apply(lambda x: float(x.split("dur")[1][0]) if "dur" in x else 3)

cat['formant'] = cat['filename'].apply(lambda x: float(x.split("F")[1][0]) if "F" in x else 3)
cat['spectilt'] = cat['filename'].apply(lambda x: float(x.split("st")[1][0]) if "st" in x else 3)
cat['duration'] = cat['filename'].apply(lambda x: float(x.split("dur")[1][0]) if "dur" in x else 3)

eng['formant'] = eng['filename'].apply(lambda x: float(x.split("F")[1][0]) if "F" in x else 3)
eng['spectilt'] = eng['filename'].apply(lambda x: float(x.split("st")[1][0]) if "st" in x else 3)
eng['duration'] = eng['filename'].apply(lambda x: float(x.split("dur")[1][0]) if "dur" in x else 3)

spa['formant_center'] = spa['formant'].apply(lambda x: x-3)
spa['spectilt_center'] = spa['spectilt'].apply(lambda x: x-3)
spa['duration_center'] = spa['duration'].apply(lambda x: x-3)

cat['formant_center'] = cat['formant'].apply(lambda x: x-3)
cat['spectilt_center'] = cat['spectilt'].apply(lambda x: x-3)
cat['duration_center'] = cat['duration'].apply(lambda x: x-3)

eng['formant_center'] = eng['formant'].apply(lambda x: x-3)
eng['spectilt_center'] = eng['spectilt'].apply(lambda x: x-3)
eng['duration_center'] = eng['duration'].apply(lambda x: x-3)

cat_rt = cat[(cat['rt']>200) & (cat['rt']<2500)].copy()
spa_rt = spa[(spa['rt']>200) & (spa['rt']<2500)].copy()
eng_rt = eng[(eng['rt']>200) & (eng['rt']<2500)].copy()
print("init cat length:", len(cat), "final cat len:", len(cat_rt))
print("init spa length:", len(spa), "final spa len:", len(spa_rt))
print("init eng length:", len(eng), "final eng len:", len(eng_rt))

cat_later = cat_rt[cat_rt['block']!=0].copy()
spa_later = spa_rt[spa_rt['block']!=0].copy()
eng_later = eng_rt[eng_rt['block']!=0].copy()
print("init cat length:", len(cat_rt), "final cat len:", len(cat_later))
print("init spa length:", len(spa_rt), "final spa len:", len(spa_later))
print("init eng length:", len(eng_rt), "final eng len:", len(eng_later))

## Adding BLP info to data

Use `03_blp_pca` in the blp folder to prepare the components from the blp data for each participant.

spa_later = spa_later.merge(parts, on="partID", how = "left")
cat_later = cat_later.merge(parts, on="partID", how = "left")
eng_later = eng_later.merge(parts, on="partID", how = "left")
spa_later.head(2)

blp_pca = pd.read_csv("../blp/data/pca_blp.csv")
blp_pca.head()

spa_later = spa_later.merge(blp_pca, on=["partID", 'lang_profile'], how = "left")
cat_later = cat_later.merge(blp_pca, on=["partID", 'lang_profile'], how = "left")
eng_later = eng_later.merge(blp_pca, on=["partID", 'lang_profile'], how = "left")

spa_later.to_csv("data/span_r.csv", index = None)
cat_later.to_csv("data/cat_r.csv", index = None)
eng_later.to_csv("data/eng_r.csv", index = None)
