==> perception/2_logit.Rmd <==
---
title: "Logit models of perception data"
author: "Annie Helms (annie_helms@berkeley.edu)"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, dev = "cairo_pdf")
options(scipen = 999) # avoid scientific notation
```

Library packages

```{r message = FALSE, warning = FALSE}
library(lmerTest) #mixed effects linear regression
library(dplyr)
library(ggplot2)
library(Cairo)
library(effects)
library(ragg)
library(gridExtra)
library(ggside)
library(sjPlot)
library(emmeans)
library(ggstatsplot)
library(pscl)
library(ggtext)
library(kableExtra) # formats dataframes in html output
```

Set working directory.

```{r message = FALSE, echo = FALSE}
setwd("C:/Users/atarv/Documents/github/diss/perception/")
```


## First read in Catalan perception data

`Response_binary` is coded where "ult" = ultimate stress; "pen" = penultimate stress, so the estimates are leveled according to 'ult' responses.
```{r}
# function to save warning
myTryCatch <- function(expr) {
  warn <- err <- NULL
  value <- withCallingHandlers(
    tryCatch(expr, error=function(e) {
      err <<- e
      NULL
    }), warning=function(w) {
      warn <<- w
      invokeRestart("muffleWarning")
    })
  list(value=value, warning=warn, error=err)
}
```

```{r}
options(width = 10000)
data = read.csv("data/cat_r.csv", 
                  stringsAsFactors = TRUE)
later_blocks = data %>%
  filter(block!=0)

model_lang_plot = glm(response_binary ~ formant_center*duration_center*spectilt_center*pc1 +
                     formant_center*duration_center*spectilt_center*pc2, data = later_blocks, family = "binomial")
model_lang = glmer(response_binary ~ formant_center*duration_center*spectilt_center*pc1 +
                     formant_center*duration_center*spectilt_center*pc2 + (1|partID), data = later_blocks, family = "binomial")
summary(model_lang)

# get list of all participants
parts_df= read.csv("../blp/data/parts.csv", 
                  stringsAsFactors = TRUE)
parts = unique(parts_df$partID)

# for loop that will calculate estimate for each participant for each acoustic measure and output df
warning_list = c()
vowels_list = c()
vowels_std = c()
st_list = c()
st_std = c()
dur_list = c()
dur_std = c()
parts_list = c()
#lang_profile_list = c()

for (i in parts){
  if (i %in% later_blocks$partID) {
    data = later_blocks %>%
      filter(partID == i)
    warn_found = as.character(myTryCatch(glm(response_binary ~ formant_center*duration_center*spectilt_center,
                                             data = data, family = "binomial"))[2])
    if (warn_found == "NULL") {
      warning_list = append(warning_list, "0")
    } else {
      warning_list = append(warning_list, warn_found)
    }
    model = glm(response_binary ~ formant_center*duration_center*spectilt_center, data = data, family = "binomial")
    slope_v = summary(model)$coefficients[2,1]
    std_v = summary(model)$coefficients[2,2]
    slope_dur = summary(model)$coefficients[3,1]
    std_dur = summary(model)$coefficients[3,2]
    slope_st = summary(model)$coefficients[4,1]
    std_st = summary(model)$coefficients[4,2]
    vowels_list = c(vowels_list, slope_v)
    vowels_std = c(vowels_std, std_v)
    parts_list = append(parts_list, i)
    dur_list = c(dur_list, slope_dur)
    dur_std = c(dur_std, std_dur)
    st_list = c(st_list, slope_st)
    st_std = c(st_std, std_st)
    }
}

df_cat = data.frame(parts_list, #lang_profile_list,
                    warning_list, vowels_list, vowels_std, dur_list, dur_std, st_list, st_std)
df_cat = df_cat %>% 
        rename("partID" = "parts_list",
               #"lang_profile" = "lang_profile_list",
               "cat_warning" = "warning_list",
               "cat_vowel_estimate" = "vowels_list",
               "cat_vowel_std" = "vowels_std",
               "cat_duration_estimate" = "dur_list",
               "cat_duration_std" = "dur_std",
               "cat_spectilt_estimate" = "st_list",
               "cat_spectilt_std" = "st_std")
head(df_cat)
```

Visualizations of significant interactions

```{r}
set_theme(base = theme_bw())
cat_vowel_pc1 = plot_model(model_lang_plot, type = "pred", terms = c("pc1 [all]", "formant_center"), axis.title = c("Principal Component 1", "% ponDIS"),
           title = "Catalan perception - Vowel quality", legend.title = "Steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "none") +
  scale_y_continuous(labels=c(0,25,50,75,100), breaks=c(0,0.25, 0.5, 0.75, 1), limits = c(0,1))
cat_vowel_pc1
```
```{r}
cairo_pdf("figures/cat_vowel_pc1.pdf", family = "serif", width = 4, height = 4)
cat_vowel_pc1
dev.off()
```



```{r}
set_theme(base = theme_bw())
cat_dur_pc1 = plot_model(model_lang_plot, type = "pred", terms = c("pc1 [all]", "duration_center"), axis.title = c("Principal Component 1", "% ponDIS"),
           title = "Catalan perception - Duration",
           legend.title = "Steps",
           pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "none") +
  scale_y_continuous(labels=c(0,25, 50, 75, 100), breaks=c(0,0.25, 0.5, 0.75, 1), limits = c(0,1))
cat_dur_pc1
```

```{r}
cairo_pdf("figures/cat_dur_pc1.pdf", family = "serif", width = 4, height = 4)
cat_dur_pc1
dev.off()
```


```{r}
set_theme(base = theme_bw())
cat_st_pc1 = plot_model(model_lang_plot, type = "pred", terms = c("pc1 [all]", "spectilt_center"), axis.title = c("Principal Component 1", "% ponDIS"),
           title = "Catalan perception - Spectral tilt", legend.title = "Steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = c(0.5,0.1), legend.direction="horizontal") +
  scale_y_continuous(labels=c(0,25, 50, 75, 100), breaks=c(0,0.25, 0.5, 0.75, 1), limits = c(0,1))
cat_st_pc1
```
```{r}
cairo_pdf("figures/cat_st_pc1.pdf", family = "serif", width = 4, height = 4)
cat_st_pc1
dev.off()
```
```{r}
cairo_pdf("figures/cat_pc1.pdf", family = "serif", width = 12, height = 4)
grid.arrange(cat_vowel_pc1, cat_st_pc1, cat_dur_pc1, ncol = 3)
dev.off()
```


```{r}
set_theme(base = theme_bw())
cat_vowel_pc2 = plot_model(model_lang_plot, type = "pred", terms = c("pc2 [all]", "formant_center"), axis.title = c("Principal Component 2", "% ponDIS"),
           title = "Catalan perception - Vowel quality", legend.title = "Steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "none", legend.direction="horizontal") +
  scale_y_continuous(labels=c(25, 50, 75, 100), breaks=c(0.25, 0.5, 0.75, 1), limits = c(0,1))
cat_vowel_pc2
```
```{r}
cairo_pdf("figures/cat_vowel_pc2.pdf", family = "serif", width = 8, height = 4)
cat_vowel_pc2
dev.off()
```

```{r}
set_theme(base = theme_bw())
cat_st_pc2 = plot_model(model_lang_plot, type = "pred", terms = c("pc2 [all]", "spectilt_center"), axis.title = c("Principal Component 2", "% ponDIS"),
           title = "Catalan perception - Spectral tilt", legend.title = "Steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = c(0.5, 0.08), legend.direction = "horizontal") +
  scale_y_continuous(labels=c(25, 50, 75, 100), breaks=c(0.25, 0.5, 0.75, 1), limits = c(0,1))
cat_st_pc2
```
```{r}
cairo_pdf("figures/cat_st_pc2.pdf", family = "serif", width = 8, height = 4)
cat_st_pc2
dev.off()
```

```{r}
cairo_pdf("figures/cat_pc2.pdf", family = "serif", width = 8, height = 4)
grid.arrange(cat_vowel_pc2, cat_st_pc2, ncol = 2)
dev.off()
```

```{r}
set_theme(base = theme_bw())
cat_vowel_st_pc2 = plot_model(model_lang_plot, type = "pred", terms = c("pc2 [all]", "formant_center", "spectilt_center"), axis.title = c("Principal Component 2", "% ponDIS"),
           title = "Catalan perception", legend.title = "Vowel quality steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")

cat_vowel_st_pc2$facet$params$nrow=1

cat_vowel_st_pc2$data$facet <- ifelse(cat_vowel_st_pc2$data$facet == "spectilt_center = -2", "Spectral tilt step 1",
ifelse(cat_vowel_st_pc2$data$facet == "spectilt_center = -1", "Spectral tilt step 2",
ifelse(cat_vowel_st_pc2$data$facet == "spectilt_center = 0", "Spectral tilt step 3",
ifelse(cat_vowel_st_pc2$data$facet == "spectilt_center = 1", "Spectral tilt step 4",
ifelse(cat_vowel_st_pc2$data$facet == "spectilt_center = 2", "Spectral tilt step 5", 99)))))

cat_vowel_st_pc2
```

```{r}
cairo_pdf("figures/cat_vowel_st_pc2.pdf", family = "serif", width = 10, height = 4)
cat_vowel_st_pc2
dev.off()
```

## Now read in Spanish data

```{r}
options(width = 10000)
data = read.csv("data/span_r.csv", 
                  stringsAsFactors = TRUE)
later_blocks = data %>%
  filter(block!=0)

model_lang_plot = glm(response_binary ~ formant_center*duration_center*spectilt_center*pc1 + 
                     formant_center*duration_center*spectilt_center*pc2, data = later_blocks, family = "binomial")
model_lang = glmer(response_binary ~ formant_center*duration_center*spectilt_center*pc1 + 
                     formant_center*duration_center*spectilt_center*pc2 + (1|partID), data = later_blocks, family = "binomial")
summary(model_lang)

# get list of all participants
parts_df= read.csv("../blp/data/parts.csv", 
                  stringsAsFactors = TRUE)
parts = unique(parts_df$partID)

# for loop that will calculate estimate for each participant for each acoustic measure and output df
warning_list = c()
vowels_list = c()
vowels_std = c()
st_list = c()
st_std = c()
dur_list = c()
dur_std = c()
parts_list = c()
#lang_profile_list = c()

for (i in parts){
  if (i %in% later_blocks$partID) {
    data = later_blocks %>%
      filter(partID == i)
    warn_found = as.character(myTryCatch(glm(response_binary ~ formant_center*duration_center*spectilt_center,
                                             data = data, family = "binomial"))[2])
    if (warn_found == "NULL") {
      warning_list = append(warning_list, "0")
    } else {
      warning_list = append(warning_list, warn_found)
    }
    model = glm(response_binary ~ formant_center*duration_center*spectilt_center, data = data, family = "binomial")
    slope_v = summary(model)$coefficients[2,1]
    std_v = summary(model)$coefficients[2,2]
    slope_dur = summary(model)$coefficients[3,1]
    std_dur = summary(model)$coefficients[3,2]
    slope_st = summary(model)$coefficients[4,1]
    std_st = summary(model)$coefficients[4,2]
    vowels_list = c(vowels_list, slope_v)
    vowels_std = c(vowels_std, std_v)
    parts_list = append(parts_list, i)
    dur_list = c(dur_list, slope_dur)
    dur_std = c(dur_std, std_dur)
    st_list = c(st_list, slope_st)
    st_std = c(st_std, std_st)
    }
}


df_spa = data.frame(parts_list, #lang_profile_list,
                    warning_list,
                vowels_list, vowels_std, dur_list, dur_std, st_list, st_std)
df_spa = df_spa %>% 
        rename("partID" = "parts_list",
               #"lang_profile" = "lang_profile_list",
               "spa_warning" = "warning_list",
               "spa_vowel_estimate" = "vowels_list",
               "spa_vowel_std" = "vowels_std",
               "spa_duration_estimate" = "dur_list",
               "spa_duration_std" = "dur_std",
               "spa_spectilt_estimate" = "st_list",
               "spa_spectilt_std" = "st_std")
head(df_spa)
```

Visualizations of significant interactions

```{r}
set_theme(base = theme_bw())
spa_vowel_pc1 = plot_model(model_lang_plot, type = "pred", terms = c("pc1 [all]", "formant_center"), axis.title = c("Principal Component 1", "% ponDIS"),
           title = "Spanish perception", legend.title = "Vowel quality steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")
spa_vowel_pc1
```
```{r}
cairo_pdf("figures/spa_vowel_pc1.pdf", family = "serif", width = 8, height = 4)
spa_vowel_pc1
dev.off()
```



## Now read in English data for a comparison with the monolingual speakers

```{r}
options(width = 10000)
# read in data
eng = read.csv("data/eng_r.csv", 
                  stringsAsFactors = TRUE)

later_blocks = eng %>%
  filter(block!=0)

model_lang_plot = glm(response_binary ~ formant_center*duration_center*spectilt_center*lang_profile, data = later_blocks, family = "binomial")
model_lang = glmer(response_binary ~ formant_center*duration_center*spectilt_center*lang_profile + (1|partID),
                   data = later_blocks, family = "binomial")
summary(model_lang)

# get list of all participants
parts_df= read.csv("../blp/data/parts.csv", 
                  stringsAsFactors = TRUE)
parts = unique(parts_df$partID)

# for loop that will calculate estimate for each participant for each acoustic measure and output df
warning_list = c()
vowels_list = c()
vowels_std = c()
st_list = c()
st_std = c()
dur_list = c()
dur_std = c()
parts_list = c()
#lang_profile_list = c()

for (i in parts){
  if (i %in% later_blocks$partID) {
    data = later_blocks %>%
      filter(partID == i)
    warn_found = as.character(myTryCatch(glm(response_binary ~ formant_center*duration_center*spectilt_center,
                                             data = data, family = "binomial"))[2])
    if (warn_found == "NULL") {
      warning_list = append(warning_list, "0")
    } else {
      warning_list = append(warning_list, warn_found)
    }
    model = glm(response_binary ~ formant_center*duration_center*spectilt_center, data = data, family = "binomial")
    slope_v = summary(model)$coefficients[2,1]
    std_v = summary(model)$coefficients[2,2]
    slope_dur = summary(model)$coefficients[3,1]
    std_dur = summary(model)$coefficients[3,2]
    slope_st = summary(model)$coefficients[4,1]
    std_st = summary(model)$coefficients[4,2]
    vowels_list = c(vowels_list, slope_v)
    vowels_std = c(vowels_std, std_v)
    parts_list = append(parts_list, i)
    dur_list = c(dur_list, slope_dur)
    dur_std = c(dur_std, std_dur)
    st_list = c(st_list, slope_st)
    st_std = c(st_std, std_st)
    }
}

df_eng = data.frame(parts_list, #lang_profile_list,
                    warning_list,
                vowels_list, vowels_std, dur_list, dur_std, st_list, st_std)
df_eng = df_eng %>% 
        rename("partID" = "parts_list",
               #"lang_profile" = "lang_profile_list",
               "eng_warning" = "warning_list",
               "eng_vowel_estimate" = "vowels_list",
               "eng_vowel_std" = "vowels_std",
               "eng_duration_estimate" = "dur_list",
               "eng_duration_std" = "dur_std",
               "eng_spectilt_estimate" = "st_list",
               "eng_spectilt_std" = "st_std")
head(df_eng)
```

Post-hoc for language profile interaction:

```{r}
emmip(model_lang, formant_center ~ lang_profile, cov.reduce = range)
emt <- emtrends(model_lang, "lang_profile", var = "formant_center")
emt
pairs(emt)
```


```{r}
emtrends(
  model_lang,
  pairwise ~ lang_profile | duration_center, var = "formant_center",
  at = list(duration_center = seq(-2, 2, by = 1))
)
```
```{r}
emtrends(
  model_lang,
   ~ formant_center | duration_center, var = "lang_profile",
  at = list(formant_center = seq(-2, 2, by = 1))
)
```


Visualizations of significant interactions

```{r}
set_theme(base = theme_bw())
eng_vowel_lang = plot_model(model_lang_plot, type = "pred", terms = c("formant_center", "lang_profile"),
                            axis.title = c("Vowel quality steps", "% ponDIS"),
           title = "English perception",
           legend.title = "Language Profile",
           pred.type = "re") +
  scale_color_discrete(labels=c('L1 English', 'L3 English', 'Monolingual')) +
  scale_x_continuous(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")
eng_vowel_lang
```

```{r}
cairo_pdf("figures/eng_vowel_lang.pdf", family = "serif", width = 8, height = 4)
eng_vowel_lang
dev.off()
```

```{r}
emmeans(model_lang_plot, list(pairwise ~ lang_profile*formant_center), adjust="tukey")
ems = emmeans(model_lang_plot, c("lang_profile", "formant_center"), infer = c(T, T))
emmeans::eff_size(ems, sigma = sigma(model_lang_plot), edf = df.residual(model_lang_plot))
```


```{r}
set_theme(base = theme_bw())
eng_vowel_dur_lang = plot_model(model_lang_plot, type = "pred", terms = c("formant_center", "lang_profile", "duration_center"), axis.title = c("Vowel quality steps", "% ponDIS"),
           title = "English perception", legend.title = "Language Profile", pred.type = "re") +
  scale_color_discrete(labels=c('L1 English', 'L3 English', 'Monolingual')) +
  scale_x_continuous(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")

eng_vowel_dur_lang$facet$params$nrow=1

eng_vowel_dur_lang$data$facet <- ifelse(eng_vowel_dur_lang$data$facet == "duration_center = -2", "Duration step 1",
ifelse(eng_vowel_dur_lang$data$facet == "duration_center = -1", "Duration step 2",
ifelse(eng_vowel_dur_lang$data$facet == "duration_center = 0", "Duration step 3",
ifelse(eng_vowel_dur_lang$data$facet == "duration_center = 1", "Duration step 4",
ifelse(eng_vowel_dur_lang$data$facet == "duration_center = 2", "Duration step 5", 99)))))

eng_vowel_dur_lang
```

```{r}
cairo_pdf("figures/eng_vowel_dur_lang.pdf", family = "serif", width = 10, height = 4)
eng_vowel_dur_lang
dev.off()
```

```{r}
set_theme(base = theme_bw())
eng_vowel_st_lang = plot_model(model_lang_plot, type = "pred", terms = c("formant_center", "lang_profile", "spectilt_center"), axis.title = c("Vowel quality steps", "% ponDIS"),
           title = "English perception", legend.title = "Language Profile", pred.type = "re") +
  scale_color_discrete(labels=c('L1 English', 'L3 English', 'Monolingual')) +
  scale_x_continuous(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")

eng_vowel_st_lang$facet$params$nrow=1

eng_vowel_st_lang$data$facet <- ifelse(eng_vowel_st_lang$data$facet == "spectilt_center = -2", "Spectral tilt step 1",
ifelse(eng_vowel_st_lang$data$facet == "spectilt_center = -1", "Spectral tilt step 2",
ifelse(eng_vowel_st_lang$data$facet == "spectilt_center = 0", "Spectral tilt step 3",
ifelse(eng_vowel_st_lang$data$facet == "spectilt_center = 1", "Spectral tilt step 4",
ifelse(eng_vowel_st_lang$data$facet == "spectilt_center = 2", "Spectral tilt step 5", 99)))))

eng_vowel_st_lang
```
```{r}
cairo_pdf("figures/eng_vowel_st_lang.pdf", family = "serif", width = 10, height = 4)
eng_vowel_st_lang
dev.off()
```



## Now look at English perception just across the trilingual speakers

```{r}
options(width = 10000)
# read in data
eng_tri = eng %>%
  filter(block!=0, lang_profile!= "mono")

model_lang_plot = glm(response_binary ~ formant_center*duration_center*spectilt_center*pc1 +
                        formant_center*duration_center*spectilt_center*pc2, data = eng_tri, family = "binomial")
model_lang = glmer(response_binary ~ formant_center*duration_center*spectilt_center*pc1 +
                        formant_center*duration_center*spectilt_center*pc2 + (1|partID), data = eng_tri, family = "binomial")
summary(model_lang)
```

Visualizations for significant interactions:

```{r}
set_theme(base = theme_bw())
eng_vowel_pc1 = plot_model(model_lang_plot, type = "pred", terms = c("pc1 [all]", "formant_center"), axis.title = c("Principal Component 1", "% ponDIS"),
           title = "English perception", legend.title = "Vowel quality steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")
eng_vowel_pc1
```
```{r}
cairo_pdf("figures/eng_vowel_pc1.pdf", family = "serif", width = 8, height = 4)
eng_vowel_pc1
dev.off()
```

```{r}
set_theme(base = theme_bw())
eng_dur_pc2 = plot_model(model_lang_plot, type = "pred", terms = c("pc2 [all]", "duration_center"), axis.title = c("Principal Component 2", "% ponDIS"),
           title = "English perception", legend.title = "Duration steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")
eng_dur_pc2
```
```{r}
cairo_pdf("figures/eng_dur_pc2.pdf", family = "serif", width = 8, height = 4)
eng_dur_pc2
dev.off()
```
```{r}
set_theme(base = theme_bw())
eng_vowel_st_pc2 = plot_model(model_lang_plot, type = "pred", terms = c("pc2 [all]", "formant_center", "spectilt_center"), axis.title = c("Principal Component 2", "% ponDIS"),
           title = "English perception", legend.title = "Vowel quality steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")

eng_vowel_st_pc2$facet$params$nrow=1

eng_vowel_st_pc2$data$facet <- ifelse(eng_vowel_st_pc2$data$facet == "spectilt_center = -2", "Spectral tilt step 1",
ifelse(eng_vowel_st_pc2$data$facet == "spectilt_center = -1", "Spectral tilt step 2",
ifelse(eng_vowel_st_pc2$data$facet == "spectilt_center = 0", "Spectral tilt step 3",
ifelse(eng_vowel_st_pc2$data$facet == "spectilt_center = 1", "Spectral tilt step 4",
ifelse(eng_vowel_st_pc2$data$facet == "spectilt_center = 2", "Spectral tilt step 5", 99)))))

eng_vowel_st_pc2
```
```{r}
cairo_pdf("figures/eng_vowel_st_pc2.pdf", family = "serif", width = 8, height = 4)
eng_vowel_st_pc2
dev.off()
```

## Now combine into larger df

```{r}
df_all = merge(df_cat, df_spa, by = "partID", all.x = TRUE, all.y = TRUE)
df_all = merge(df_all, df_eng, by = "partID", all.x = TRUE, all.y = TRUE)
sample(df_all)
```

## Now combine with pca and lang_profile
```{r}
lang_info = eng %>%
  group_by(partID) %>%
  select(partID, lang_profile, pc1, pc2) %>%
  unique()

df_blp = merge(df_all, lang_info, by = "partID")
write.csv(df_blp, "data/estimates_and_blp.csv", row.names = FALSE)
head(df_blp)
```

## Now get English 3-way interaction cleared up

```{r}
eng = read.csv("data/eng_r.csv", 
                  stringsAsFactors = TRUE)
later_blocks = eng %>%
  filter(block!=0 & lang_profile=="mono")

model_lang_plot = glm(response_binary ~ formant_center*duration_center*spectilt_center, data = later_blocks, family = "binomial")
model_lang = glmer(response_binary ~ formant_center*duration_center*spectilt_center + (1|partID), data = later_blocks, family = "binomial")
summary(model_lang)
```
```{r}
set_theme(base = theme_bw())
eng_dur_vowel_mono = plot_model(model_lang_plot, type = "pred", terms = c("formant_center", "duration_center"), axis.title = c("Formant steps", "% ponDIS"),
           title = "English perception", legend.title = "Duration steps", pred.type = "re") +
  scale_color_discrete(labels=c('1', '2', '3', '4', '5')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")
eng_dur_vowel_mono
```



```{r}
eng = read.csv("data/eng_r.csv", 
                  stringsAsFactors = TRUE)
later_blocks_l1 = eng %>%
  filter(block!=0 & lang_profile=="l1_eng")

model_lang_plot = glm(response_binary ~ formant_center*duration_center*spectilt_center, data = later_blocks_l1, family = "binomial")
model_lang = glmer(response_binary ~ formant_center*duration_center*spectilt_center + (1|partID), data = later_blocks_l1, family = "binomial")
summary(model_lang)
```

```{r}
eng = read.csv("data/eng_r.csv", 
                  stringsAsFactors = TRUE)
later_blocks_l3 = eng %>%
  filter(block!=0 & lang_profile=="l3_eng")

model_lang_plot = glm(response_binary ~ formant_center*duration_center*spectilt_center, data = later_blocks_l3, family = "binomial")
model_lang = glmer(response_binary ~ formant_center*duration_center*spectilt_center + (1|partID), data = later_blocks_l3, family = "binomial")
summary(model_lang)
```
