==> perception/8_eng_all_bayes.Rmd <==
---
title: "Logit models of perception data with Bayesian predictive modelling"
author: "Annie Helms (annie_helms@berkeley.edu)"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, dev = "cairo_pdf", out.height = 600)
options(scipen = 999) # avoid scientific notation
```

Library packages

```{r message = FALSE, warning = FALSE}
library(lmerTest) #mixed effects linear regression
library(dplyr)
library(ggplot2)
library(Cairo)
library(effects)
library(ragg)
library(gridExtra)
library(grid)
library(ggside)
library(sjPlot)
library(emmeans)
library(ggstatsplot)
library(pscl)
library(bayesplot)
library(brms)
library(ggtext)
library(kableExtra) # formats dataframes in html output
```

Set working directory.

```{r message = FALSE, echo = FALSE}
setwd("C:/Users/atarv/Documents/github/diss/perception/")
```


```{r}
options(width = 10000)
df = read.csv("data/eng_r.csv", 
                  stringsAsFactors = TRUE)
head(df)
```

## English perception

```{r}
# binarized label for judgment
data = df %>%
  mutate(response_onehot = case_when(response_binary=="pen"~0,
                                    response_binary=="ult"~1),
         formant_jit = jitter(formant_center, amount=.001),
         spectilt_jit = jitter(spectilt_center, amount=.001),
         duration_jit = jitter(duration_center, amount=.001))
head(data)
```


```{r}
formula = "response_onehot ~ formant_jit*spectilt_jit*duration_jit*lang_profile + (formant_jit*spectilt_jit*duration_jit|partID)"

priors <- get_prior(formula,data,family=bernoulli())

model <- brm(formula=bf(formula,autocor = ~ 0),data=data, family=bernoulli(), 
            prior=priors, cores = 8, save_pars = save_pars(all = TRUE), iter=4000)
```

Save or load model

```{r}
#save(model, file="data/eng_all_model.Rdata")
load("data/eng_all_model.Rdata")
```

Examine results

```{r}
summary(model)
bayes_R2(model)
```

Visualize results

```{r}
posterior <- as.array(model)  # prepare to plot the brm result
dim(posterior)
dimnames(posterior)$variable[1:50]
```

```{r}
coeff_plot = mcmc_areas(posterior, pars = c("b_formant_jit","b_spectilt_jit","b_duration_jit","b_lang_profilel3_eng", "b_lang_profilemono", "b_formant_jit:spectilt_jit", "b_formant_jit:duration_jit", "b_spectilt_jit:duration_jit", "b_formant_jit:lang_profilel3_eng", "b_formant_jit:lang_profilemono", "b_spectilt_jit:lang_profilel3_eng", "b_spectilt_jit:lang_profilemono", "b_duration_jit:lang_profilel3_eng", "b_duration_jit:lang_profilemono", "b_formant_jit:spectilt_jit:duration_jit", "b_formant_jit:spectilt_jit:lang_profilel3_eng", "b_formant_jit:spectilt_jit:lang_profilemono", "b_formant_jit:duration_jit:lang_profilel3_eng", "b_formant_jit:duration_jit:lang_profilemono", "b_spectilt_jit:duration_jit:lang_profilel3_eng", "b_spectilt_jit:duration_jit:lang_profilemono", "b_formant_jit:spectilt_jit:duration_jit:lang_profilel3_eng", "b_formant_jit:spectilt_jit:duration_jit:lang_profilemono"), 
          prob = 0.8, # 80% intervals
          prob_outer = 0.95, # 95%
          point_est = "median") +
  xlim(-2,2)+
scale_y_discrete(labels=c("Vowel quality:Spectral tilt:Duration:Monolingual", "Vowel quality:Spectral tilt:Duration:L3 English", "Spectral tilt:Duration:Monolingual", "Spectral tilt:Duration:L3 English", "Vowel quality:Duration:Monolingual", "Vowel quality:Duration:L3 English", "Vowel quality:Spectral tilt:Monolingual", "Vowel quality:Spectral tilt:L3 English", "Vowel quality:Spectral tilt:Duration", "Duration:Monolingual", "Duration:L3 English", "Spectral tilt:Monolingual", "Spectral tilt:L3 English", "Vowel quality:Monolingual", "Vowel quality:L3 English", "Spectral tilt:Duration", "Vowel quality:Duration", "Vowel quality:Spectral tilt", "Monolingual", "L3 English", "Duration", "Spectral tilt", "Vowel quality"), limits=rev) +
  labs(title = "Coefficients of English perception", x ="Coefficient, sd(Oxtyone Stress)") +theme(plot.title=element_text(hjust = 0.5))
coeff_plot
```


```{r}
cairo_pdf("figures/bayes/eng_lang_perc_coeff.pdf", family = "serif", width = 8, height = 6)
coeff_plot
dev.off()
```

```{r}
eng_vow_lang = conditional_effects(model, effects="formant_jit:lang_profile", init_conditions = "lang_profile")
```

```{r}
emt <- emtrends(model, "lang_profile", var = "formant_jit")
emt
pairs(emt)
```

```{r}
eng_vow_lang_plot = plot(eng_vow_lang, plot = FALSE)[[1]] + labs(title = "English perception",
                                                                         x = "Vowel quality steps",
                                                                         y = "Proportion ponDIS")+
guides(color=guide_legend(title="Language Profile"), fill = "none")+
  scale_color_discrete(breaks = c("l1_eng", 'l3_eng', 'mono'),
                       labels=c('L1 English', 'L3 English', 'Monolingual')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")

# eng_vow_dur_lang_plot$data$cond__ <- ifelse(eng_vow_dur_lang_plot$data$cond__ == "1", "Duration step -2",
#   ifelse(eng_vow_dur_lang_plot$data$cond__ == "2", "Duration step -1",
#   ifelse(eng_vow_dur_lang_plot$data$cond__ == "3", "Duration step 0",
#   ifelse(eng_vow_dur_lang_plot$data$cond__ == "4", "Duration step 1",
#   ifelse(eng_vow_dur_lang_plot$data$cond__ == "5", "Duration step 2", 99)))))
# eng_vow_dur_lang_plot$facet$params$ncol=5
# 
# eng_vow_lang_plot$data$cond__ <- factor(eng_vow_dur_lang_plot$data$cond__, levels = c("Duration step -2", "Duration step -1",
#                                                                                           "Duration step 0", "Duration step 1",
#                                                                                           "Duration step 2"))

eng_vow_lang_plot
```

```{r}
cairo_pdf("figures/bayes/eng_lang_vowel.pdf", family = "serif", width = 6, height = 4)
eng_vow_lang_plot
dev.off()
```


```{r}
conditions_lang = data.frame(duration_jit = c(-2, -1, 0, 1, 2))
eng_vow_dur_lang = conditional_effects(model, effects = "formant_jit:lang_profile", conditions = conditions_lang)
```

```{r}
eng_vow_dur_lang_plot = plot(eng_vow_dur_lang, plot = FALSE)[[1]] + labs(title = "English perception",
                                                                         x = "Vowel quality steps",
                                                                         y = "Proportion ponDIS") + 
guides(color=guide_legend(title="Language Profile"), fill = 'none') +
  scale_color_discrete(breaks = c("l1_eng", 'l3_eng', 'mono'),
                       labels=c('L1 English', 'L3 English', 'Monolingual')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")

eng_vow_dur_lang_plot$data$cond__ <- ifelse(eng_vow_dur_lang_plot$data$cond__ == "1", "Duration step -2",
  ifelse(eng_vow_dur_lang_plot$data$cond__ == "2", "Duration step -1",
  ifelse(eng_vow_dur_lang_plot$data$cond__ == "3", "Duration step 0",
  ifelse(eng_vow_dur_lang_plot$data$cond__ == "4", "Duration step 1",
  ifelse(eng_vow_dur_lang_plot$data$cond__ == "5", "Duration step 2", 99)))))
eng_vow_dur_lang_plot$facet$params$ncol=5

eng_vow_dur_lang_plot$data$cond__ <- factor(eng_vow_dur_lang_plot$data$cond__, levels = c("Duration step -2", "Duration step -1",
                                                                                          "Duration step 0", "Duration step 1",
                                                                                          "Duration step 2"))

eng_vow_dur_lang_plot
```

```{r}
cairo_pdf("figures/bayes/eng_lang_vowel_dur.pdf", family = "serif", width = 8, height = 4)
eng_vow_dur_lang_plot
dev.off()
```

```{r}
conditions_lang = data.frame(spectilt_jit = c(-2, -1, 0, 1, 2))
eng_vow_st_lang = conditional_effects(model, effects = "formant_jit:lang_profile", conditions = conditions_lang)
```

```{r}
eng_vow_st_lang_plot = plot(eng_vow_st_lang, plot = FALSE)[[1]] + labs(title = "English perception",
                                                                         x = "Vowel quality steps",
                                                                         y = "Proportion ponDIS") + 
guides(color=guide_legend(title="Language Profile"), fill = 'none') +
  scale_color_discrete(breaks = c("l1_eng", 'l3_eng', 'mono'),
                       labels=c('L1 English', 'L3 English', 'Monolingual')) +
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")

eng_vow_st_lang_plot$data$cond__ <- ifelse(eng_vow_st_lang_plot$data$cond__ == "1", "Spectral tilt step -2",
  ifelse(eng_vow_st_lang_plot$data$cond__ == "2", "Spectral tilt step -1",
  ifelse(eng_vow_st_lang_plot$data$cond__ == "3", "Spectral tilt step 0",
  ifelse(eng_vow_st_lang_plot$data$cond__ == "4", "Spectral tilt step 1",
  ifelse(eng_vow_st_lang_plot$data$cond__ == "5", "Spectral tilt step 2", 99)))))
eng_vow_st_lang_plot$facet$params$ncol=5

eng_vow_st_lang_plot$data$cond__ <- factor(eng_vow_st_lang_plot$data$cond__, levels = c("Spectral tilt step -2", "Spectral tilt step -1",
                                                                                          "Spectral tilt step 0", "Spectral tilt step 1",
                                                                                          "Spectral tilt step 2"))

eng_vow_st_lang_plot
```
```{r}
cairo_pdf("figures/bayes/eng_lang_vowel_st.pdf", family = "serif", width = 8, height = 4)
eng_vow_st_lang_plot
dev.off()
```



```{r}
cat_vowel_spec_pc2_plot = plot(cat_vowel_spec_pc2, plot = FALSE)[[1]] + labs(title = "Catalan perception", x = "Vowel quality steps", y = "Proportion stressed") +
  guides(color=guide_legend(title="Spectral tilt steps"), fill = 'none') +
  scale_color_discrete(breaks = c(-1.15, 0, 1.15),
    labels=c('-1 SD', 'mean', '+1 SD')
    )+
  theme(plot.title=element_text(hjust = 0.5), legend.position = "bottom")

cat_vowel_spec_pc2_plot$data$cond__ <- ifelse(cat_vowel_spec_pc2_plot$data$cond__ == "1", "PC2 (-1 SD)",
ifelse(cat_vowel_spec_pc2_plot$data$cond__ == "2", "PC2 (mean)",
ifelse(cat_vowel_spec_pc2_plot$data$cond__ == "3", "PC2 (+1 SD)", 99)))


cat_vowel_spec_pc2_plot$data$cond__ <- factor(cat_vowel_spec_pc2_plot$data$cond__, levels = c("PC2 (-1 SD)", "PC2 (mean)", "PC2 (+1 SD)"))

cat_vowel_spec_pc2_plot
```

```{r}
cairo_pdf("figures/bayes/cat_vowel_spec_pc2_perception.pdf", family = "serif", width = 8, height = 6)
cat_vowel_spec_pc2_plot
dev.off()
```
